Bugfix
Fixed an issue where foreign keys failed to be generated as primary keys had different data types (In Reservations)

+++ parksnap_back/src/main/java/com/example/parking/entity/ReservationId.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
 @Embeddable
 public class ReservationId implements Serializable {
 
-    private Long slotId;
-    private Long userId;
-    private Long vehicleId;
+    private int slotId;
+    private int userId;
+    private int vehicleId;
 }
